+++ src/components/color/model/oklch.tsx
@@ -7,10 +7,8 @@ import { useColor } from "~/app/store";
 
 export default function InputOklch({
   onChange,
-  id,
 }: {
   onChange?: (color: OklchColorMode) => void;
-  id?: string;
 }) {
   const [{ oklch }] = useColor();
   const [color, setOklch] = useState<OklchColor>(oklch.color);
@@ -58,7 +56,7 @@ export default function InputOklch({
   }, [color, oklch]);
 
   return (
-    <div role="form" aria-label="oklch color" className="grid gap-4" id={id}>
+    <>
       <div role="none" className="relative inline-grid">
         <input
           aria-label="lightness"
@@ -125,6 +123,6 @@ export default function InputOklch({
           }}
         ></span>
       </div>
-    </div>
+    </>
   );
 }
